fat16_fseek

+++ src/fs/file.c
@@ -179,7 +179,7 @@ size_t fread(void *ptr, size_t size, size_t nmemb, int fd)
         return rc;
 }
 
-int fseek(int fd, int offset, enum file_seek_mode whence)
+int fseek(int fd, size_t offset, enum file_seek_mode whence)
 {
         struct file_descriptor *desc = file_get_descriptor(fd);
         if (!desc)
